requirements.txt
streamlit==1.18.1
tensorflow==2.16.2
keras==2.12.0
numpy==1.24.2
Pillow==9.5.0
pandas==2.0.2
